Fix selection box drawing
The XOR operator in cairo seems to be different from the XOR operator in GDK.
Use the DIFFERENCE instead to get the same behavior as before.

Also make the size of the selection rectangle an integer, so that all edges are
on a 0.5 grid. This will make sure that the lines are only 1px wide as they used
to be.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>